requirements.txt
feedparser>=6.0.10
newspaper3k>=0.2.8
transformers>=4.41.1
torch
tqdm>=4.66.4
lxml[html_clean]>=5.2.2
numpy<2.0  # Optional: to avoid compatibility issues
